requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
structlog==24.4.0
prometheus-client==0.20.0
httpx==0.27.2
PyYAML==6.0.2
tqdm==4.66.5
